+++ src/main/java/com/chainsys/newbornbabyhealthtrackingsystem/repository/HospitalStaffRepository.java
@@ -14,14 +14,14 @@ import com.chainsys.newbornbabyhealthtrackingsystem.model.Person;
  * @author shan3102
  *
  */
-public interface HospitalStaffRepository extends CrudRepository<HospitalStaff, Integer> {
-	HospitalStaff findById(int hosStaffId);
+public interface HospitalStaffRepository extends CrudRepository<HospitalStaff, Long> {
+	HospitalStaff findById(long hosStaffId);
 
 	HospitalStaff save(HospitalStaff hosStaffId);
 
-	void deleteById(int hosStaffId);
+	void deleteById(long hosStaffId);
 
 	List<HospitalStaff> findAll();
 
-//	Person findByStaffId(int staffId);
+//	Person findByStaffId(long staffId);
 }
