+++ src/main/java/com/chainsys/newbornbabyhealthtrackingsystem/services/PersonServices.java
@@ -33,25 +33,27 @@ public class PersonServices {
 		return personRepo.save(thePerson);		
 	}
 
-	public void removePerson(int personId) {
+	public void removePerson(long personId) {
 		personRepo.deleteById(personId);
 	}
 
-	public Person getPersonById(int personId) {
+	public Person getPersonById(long personId) {
 		return personRepo.findByUserId(personId);
 	}
-//	private LocationCode getPersonLocationById(int personPinId) {
-//		return locRepo.findByPinCode(personPinId);
-//	}
+//	private LocationCode getPersonLocationById(long personPinId) {
+//	return locRepo.findByPinCode(personPinId);
+//}
 
-//	@Transactional
-//	public PersonLocationDTO getPersonWithLocation(int pincode) {
-//		LocationCode loc =getPersonLocationById(pincode);
-//		PersonLocationDTO plDto= new PersonLocationDTO();
-//		plDto.setLoc(loc);
-//		return plDto;
-//	}
-	
+
+
+//@Transactional
+//public PersonLocationDTO getPersonWithLocation(long pincode) {
+//	LocationCode loc =getPersonLocationById(pincode);
+//	PersonLocationDTO plDto= new PersonLocationDTO();
+//	plDto.setLoc(loc);
+//	return plDto;
+//}
+
 //	
 
 }
